fix: exported abi utils to be used outside txRepository
