+++ src/converters/abi.converters.js
@@ -1,3 +1,5 @@
+import { removeNullFields } from '../repositories/utils'
+
 function rawAbiToEntity (data) {
   return {
     anonymous: data.anonymous,
@@ -22,4 +24,25 @@ function rawAbiInputToEntity (data) {
   }
 }
 
-export {rawAbiToEntity, rawInputToEntity, rawAbiInputToEntity}
+function abiEntityToRaw ({
+  anonymous,
+  name,
+  type,
+  abi_input: inputs
+}) {
+  const abiToReturn = {
+    anonymous,
+    name,
+    type,
+    inputs: inputs.map(({input}) => input)
+  }
+
+  return removeNullFields(abiToReturn)
+}
+
+export {
+  rawAbiToEntity,
+  rawInputToEntity,
+  rawAbiInputToEntity,
+  abiEntityToRaw
+}
